requirements.txt
APScheduler==3.11.0
fastapi==0.117.1
fitz==0.0.1.dev2
httpx==0.28.1
icalendar==6.3.1
passlib==1.7.4
Pillow==10.1.0
pydantic==1.10.6
pytesseract==0.3.13
python-dotenv==1.1.1
python_dateutil==2.8.2
python_docx==1.2.0
python_jose==3.5.0
pytz==2025.2
reportlab==4.4.3
requests==2.32.5
SQLAlchemy==2.0.43
uvicorn==0.35.0
alembic==1.16.4
